Extract SequenceNumberForIndexCallable
Extract SequenceNumberForIndexCallable from IndexUpdater to a standalone class.
